Delete radio module
