+++ esp-radio/src/wifi/os_adapter/esp32c3.rs
@@ -1,8 +1,14 @@
-use crate::hal::peripherals;
+use crate::{
+    hal::{
+        interrupt::Priority,
+        peripherals::{INTERRUPT_CORE0, WIFI},
+    },
+    sys::c_types::c_void,
+};
 
 pub(crate) fn chip_ints_on(mask: u32) {
     unsafe {
-        peripherals::INTERRUPT_CORE0::regs()
+        INTERRUPT_CORE0::regs()
             .cpu_int_enable()
             .modify(|r, w| w.bits(r.bits() | mask));
     }
@@ -10,7 +16,7 @@ pub(crate) fn chip_ints_on(mask: u32) {
 
 pub(crate) fn chip_ints_off(mask: u32) {
     unsafe {
-        peripherals::INTERRUPT_CORE0::regs()
+        INTERRUPT_CORE0::regs()
             .cpu_int_enable()
             .modify(|r, w| w.bits(r.bits() & !mask));
     }
@@ -46,11 +52,7 @@ pub(crate) unsafe extern "C" fn set_intr(
 ///   None
 ///
 /// *************************************************************************
-pub unsafe extern "C" fn set_isr(
-    n: i32,
-    f: *mut crate::sys::c_types::c_void,
-    arg: *mut crate::sys::c_types::c_void,
-) {
+pub unsafe extern "C" fn set_isr(n: i32, f: *mut c_void, arg: *mut c_void) {
     trace!("set_isr - interrupt {} function {:?} arg {:?}", n, f, arg);
     match n {
         0 => unsafe {
@@ -62,11 +64,47 @@ pub unsafe extern "C" fn set_isr(
         _ => panic!("set_isr - unsupported interrupt number {}", n),
     }
 
-    #[cfg(feature = "wifi")]
     unsafe {
-        use crate::hal::{interrupt::Priority, peripherals::WIFI};
-
         WIFI::steal().enable_mac_interrupt(Priority::Priority1);
         WIFI::steal().enable_pwr_interrupt(Priority::Priority1);
+    }
+}
+
+#[unsafe(no_mangle)]
+extern "C" fn WIFI_MAC() {
+    unsafe {
+        let (fnc, arg) = crate::wifi::ISR_INTERRUPT_1;
+
+        trace!("interrupt WIFI_MAC {:?} {:?}", fnc, arg);
+
+        if !fnc.is_null() {
+            let fnc: fn(*mut c_void) = core::mem::transmute(fnc);
+            fnc(arg);
+        }
+
+        trace!("interrupt 1 done");
+    };
+}
+
+#[unsafe(no_mangle)]
+extern "C" fn WIFI_PWR() {
+    unsafe {
+        let (fnc, arg) = crate::wifi::ISR_INTERRUPT_1;
+
+        trace!("interrupt WIFI_PWR {:?} {:?}", fnc, arg);
+
+        if !fnc.is_null() {
+            let fnc: fn(*mut c_void) = core::mem::transmute(fnc);
+            fnc(arg);
+        }
+
+        trace!("interrupt 1 done");
+    };
+}
+
+pub(crate) fn shutdown_wifi_isr() {
+    unsafe {
+        WIFI::steal().disable_mac_interrupt();
+        WIFI::steal().disable_pwr_interrupt();
     }
 }
