+++ esp-radio/src/wifi/os_adapter/esp32c5.rs
@@ -1,3 +1,8 @@
+use crate::{
+    hal::{interrupt::Priority, peripherals::WIFI},
+    sys::c_types::{c_int, c_void},
+};
+
 pub(crate) fn chip_ints_on(_mask: u32) {
     // from docs: Enable Wi-Fi interrupt / mask - No mean
     // i.e. nothing to do - we already enabled the WIFI interrupt
@@ -24,16 +29,14 @@ pub(crate) unsafe extern "C" fn set_intr(
 }
 
 pub(crate) unsafe extern "C" fn regdma_link_set_write_wait_content_dummy(
-    _arg1: *mut crate::sys::c_types::c_void,
+    _arg1: *mut c_void,
     _arg2: u32,
     _arg3: u32,
 ) {
     todo!()
 }
 
-pub(crate) unsafe extern "C" fn sleep_retention_find_link_by_id_dummy(
-    _arg1: crate::sys::c_types::c_int,
-) -> *mut crate::sys::c_types::c_void {
+pub(crate) unsafe extern "C" fn sleep_retention_find_link_by_id_dummy(_arg1: c_int) -> *mut c_void {
     todo!()
 }
 
@@ -52,11 +55,7 @@ pub(crate) unsafe extern "C" fn sleep_retention_find_link_by_id_dummy(
 ///   None
 ///
 /// *************************************************************************
-pub unsafe extern "C" fn set_isr(
-    n: i32,
-    f: *mut crate::sys::c_types::c_void,
-    arg: *mut crate::sys::c_types::c_void,
-) {
+pub unsafe extern "C" fn set_isr(n: i32, f: *mut c_void, arg: *mut c_void) {
     trace!("set_isr - interrupt {} function {:?} arg {:?}", n, f, arg);
 
     match n {
@@ -68,11 +67,48 @@ pub unsafe extern "C" fn set_isr(
         },
         _ => panic!("set_isr - unsupported interrupt number {}", n),
     }
-    #[cfg(feature = "wifi")]
-    unsafe {
-        use crate::hal::{interrupt::Priority, peripherals::WIFI};
 
+    unsafe {
         WIFI::steal().enable_mac_interrupt(Priority::Priority1);
         WIFI::steal().enable_pwr_interrupt(Priority::Priority1);
+    }
+}
+
+#[unsafe(no_mangle)]
+extern "C" fn WIFI_MAC() {
+    unsafe {
+        let (fnc, arg) = crate::wifi::ISR_INTERRUPT_1;
+
+        trace!("interrupt WIFI_MAC {:?} {:?}", fnc, arg);
+
+        if !fnc.is_null() {
+            let fnc: fn(*mut c_void) = core::mem::transmute(fnc);
+            fnc(arg);
+        }
+
+        trace!("interrupt 1 done");
+    };
+}
+
+#[unsafe(no_mangle)]
+extern "C" fn WIFI_PWR() {
+    unsafe {
+        let (fnc, arg) = crate::wifi::ISR_INTERRUPT_1;
+
+        trace!("interrupt WIFI_PWR {:?} {:?}", fnc, arg);
+
+        if !fnc.is_null() {
+            let fnc: fn(*mut c_void) = core::mem::transmute(fnc);
+            fnc(arg);
+        }
+
+        trace!("interrupt 1 done");
+    };
+}
+
+pub(crate) fn shutdown_wifi_isr() {
+    unsafe {
+        WIFI::steal().disable_mac_interrupt();
+        WIFI::steal().disable_pwr_interrupt();
     }
 }
